Cut out WatchedEvents component
